# openapi-comparison/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Library API (Simple Demo)
  version: 1.0.0
  description: API quản lý thư viện tối giản để demo tài liệu hóa và sinh code/test.
servers:
  - url: http://localhost:5000

tags:
  - name: Health
  - name: Books

paths:
  /health:
    get:
      tags: [Health]
      summary: Kiểm tra trạng thái dịch vụ
      responses:
        '200':
          description: Service healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  /books:
    get:
      tags: [Books]
      summary: Lấy danh sách sách
      responses:
        '200':
          description: Danh sách sách
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooksListResponse'
    post:
      tags: [Books]
      summary: Thêm sách mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
      responses:
        '201':
          description: Tạo thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '400':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /books/{bookId}:
    parameters:
      - $ref: '#/components/parameters/BookId'
    get:
      tags: [Books]
      summary: Lấy chi tiết sách
      responses:
        '200':
          description: Chi tiết sách
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
    put:
      tags: [Books]
      summary: Cập nhật sách
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdateRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
    delete:
      tags: [Books]
      summary: Xóa sách
      responses:
        '200':
          description: Xóa thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: Không tìm thấy sách
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    BookId:
      name: bookId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      example: 1

  schemas:
    Book:
      type: object
      required: [id, title, author]
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string

    BookCreateRequest:
      type: object
      required: [title, author]
      properties:
        title:
          type: string
          minLength: 1
        author:
          type: string
          minLength: 1

    BookUpdateRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string

    SuccessResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Success
        data:
          nullable: true

    BooksListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Book'

    BookResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Book'

    DeleteResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              nullable: true

    HealthResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                status:
                  type: string
                  example: ok

    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Book not found

    IntegerId:
      type: object
      properties:
        id:
          type: integer